requirements.txt
# requirements.txt
Django==4.2.7
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1